requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.16
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2==2.9.10
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
django-rest-auth==0.9.5
django-allauth==65.1.0
dj-rest-auth==6.0.0
drf-yasg==1.21.8
inflection==0.5.1
pytz==2024.2
pyyaml==6.0.2
uritemplate==4.1.1
